+++ packages/domain-event/src/Exception/ExceptionInterface.php
@@ -13,10 +13,6 @@ declare(strict_types=1);
 
 namespace Rekalogika\DomainEvent\Exception;
 
-class FlushNotAllowedException extends LogicException
+interface ExceptionInterface extends \Throwable
 {
-    public function __construct()
-    {
-        parent::__construct('"flush()" is not allowed inside a pre-flush domain event listener.');
-    }
 }