+++ packages/domain-event/src/Event/AbstractDomainEventDispatchEvent.php
@@ -13,9 +13,9 @@ declare(strict_types=1);
 
 namespace Rekalogika\DomainEvent\Event;
 
-class DomainEventImmediateDispatch
+abstract class AbstractDomainEventDispatchEvent
 {
-    public function __construct(private object $domainEvent)
+    final public function __construct(private object $domainEvent)
     {
     }