+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.5
 
 * fix: Properly implement `ResetInterface` on applicable services.
+* test: Add tests for Symfony 7 & PHP 8.3
 
 ## 1.2.4